.copy/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==3.0.0
mysql-connector-python==8.0.28